fix compile error when SDL is not available
Note that when using constexpr, the code inside each if branch still has to compile, regardless of the value of the constexpr.
In the case where CMake would not find SDL, the compiler would not find the declaration of SDLTerminal, since the header where it is declared is not included.